TVIST1-253: Refactored Listeners
