+++ src/Logging/EntityListener/MunicipalityListener.php
@@ -23,4 +23,23 @@ class MunicipalityListener extends AbstractEntityListener
     {
         $this->logActivity('Update', $args);
     }
+
+    /**
+     * @throws ItkDevLoggingException
+     * @throws ORMException
+     */
+    public function logActivity(string $action, LifecycleEventArgs $args): void
+    {
+        $em = $args->getEntityManager();
+
+        /** @var Municipality $municipality */
+        $municipality = $args->getObject();
+
+        foreach ($municipality->getCaseEntities() as $case) {
+            $logEntry = $this->createLogEntry($action, $case, $args);
+            $em->persist($logEntry);
+        }
+
+        $em->flush();
+    }
 }
